attendance_backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg
python-dotenv==1.0.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
bcrypt==4.1.3
mangum
psycopg2-binary
greenlet
resend